Eclipse source clean up

+++ src/main/java/com/firemerald/additionalplacements/client/gui/AbstractBetterSelectionList.java
@@ -58,19 +58,19 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 		this.minecraft = minecraft;
 		this.normalItemHeight = normalItemHeight;
 	}
-	
-	public int getX() { 
+
+	public int getX() {
 		return x;
 	}
-	
-	public int getY() { 
+
+	public int getY() {
 		return y;
 	}
-	
+
 	public int getRight() {
 		return x + width;
 	}
-	
+
 	public int getBottom() {
 		return y + height;
 	}
@@ -194,7 +194,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 
 	protected void renderDecorations(GuiGraphics guiGraphics, int mouseX, int mouseY) {
 	}
-	
+
     @Override
     public void render(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
         this.hovered = this.isMouseOver(mouseX, mouseY) ? this.getEntryAtPosition(mouseX, mouseY) : null;
@@ -251,8 +251,8 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
             resourcelocation,
             this.getX(),
             this.getY(),
-            (float)this.getRight(),
-            (float)(this.getBottom() + (int)this.getScrollAmount()),
+            this.getRight(),
+            this.getBottom() + (int)this.getScrollAmount(),
             this.getWidth(),
             this.getHeight(),
             32,
@@ -320,7 +320,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Called when a mouse button is clicked within the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the event is consumed, {@code false} otherwise.
 	 * @param mouseX the X coordinate of the mouse.
 	 * @param mouseY the Y coordinate of the mouse.
@@ -360,7 +360,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Called when a mouse button is released within the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the event is consumed, {@code false} otherwise.
 	 * @param mouseX the X coordinate of the mouse.
 	 * @param mouseY the Y coordinate of the mouse.
@@ -378,7 +378,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Called when the mouse is dragged within the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the event is consumed, {@code false} otherwise.
 	 * @param mouseX the X coordinate of the mouse.
 	 * @param mouseY the Y coordinate of the mouse.
@@ -417,7 +417,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 
 	/**
 	 * Sets the focus state of the GUI element.
-	 * 
+	 *
 	 * @param pFocused the focused GUI element.
 	 */
 	@Override
@@ -487,7 +487,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Checks if the given mouse coordinates are over the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the mouse is over the GUI element, {@code false}
 	 *         otherwise.
 	 * @param mouseX the X coordinate of the mouse.
@@ -606,7 +606,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 
 		/**
 		 * Sets the focus state of the GUI element.
-		 * 
+		 *
 		 * @param focused {@code true} to apply focus, {@code false} to remove focus
 		 */
 		@Override
@@ -631,7 +631,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 		/**
 		 * Checks if the given mouse coordinates are over the GUI element.
 		 * <p>
-		 * 
+		 *
 		 * @return {@code true} if the mouse is over the GUI element, {@code false}
 		 *         otherwise.
 		 * @param mouseX the X coordinate of the mouse.
